update test mocks
