+++ js_modules/dagster-ui/packages/ui-core/src/pipelines/__tests__/PipelineNav.test.tsx
@@ -6,7 +6,7 @@ import {TestPermissionsProvider} from '../../testing/TestPermissions';
 import {buildRepoAddress} from '../../workspace/buildRepoAddress';
 import {JobFeatureProvider} from '../JobFeatureContext';
 
-jest.mock('../../workspace/WorkspaceContext', () => ({
+jest.mock('../../../workspace/WorkspaceContext', () => ({
   ...jest.requireActual('../../workspace/WorkspaceContext'),
   useRepository: jest.fn(() => null),
 }));
