+++ js_modules/dagster-ui/packages/ui-core/src/ticks/__tests__/DryRunRequestTable.test.tsx
@@ -5,7 +5,7 @@ import {RunRequestTable} from '../DryRunRequestTable';
 import {mockRepository} from '../__fixtures__/DryRunRequestTable.fixtures';
 import {runRequests} from '../__fixtures__/SensorDryRunDialog.fixtures';
 
-jest.mock('../../workspace/WorkspaceContext', () => ({useRepository: () => mockRepository}));
+jest.mock('../../../workspace/WorkspaceContext', () => ({useRepository: () => mockRepository}));
 
 function TestComponent() {
   return (
